chore: organise popup folder files
