--- src/view/popup/__tests__/calculateEmissions.test.ts
@@ -1,29 +1,0 @@
-import { CountryName } from "../../../data/constants/CountryEmissions";
-import { calculateEmissions } from "../../../utils/emissions/calculateEmissions";
-
-const mockCountry: Map<CountryName, number> = new Map([["Australia", 0.5]]);
-
-const mockCountries: Map<CountryName, number> = new Map([
-    ["Australia", 0.2],
-    ["United Kingdom", 0.3],
-    ["Belgium", 0.2],
-    ["Bulgaria", 0.2],
-    ["Croatia", 0.1],
-]);
-
-const mockSmallBytes = 50;
-const mockLargeBytes = 500000;
-
-describe("calculateEmissions", () => {
-    it("country should return corresponding value", () => {
-        expect(calculateEmissions(mockSmallBytes, mockCountry)).toBeCloseTo(
-            0.000017695
-        );
-    });
-
-    it("multiple countries should return corresponding carbon consumption for a given transfer size", () => {
-        expect(calculateEmissions(mockLargeBytes, mockCountries)).toBeCloseTo(
-            0.16074
-        );
-    });
-});
